replaced the package-lock.json with the latest same file in main repoistory. Created a new samplepage to show SCRUM 45 homepage work, kept the old page.tsx as the same to resolve conflicts shown in github
