+++ src/app/page.tsx
@@ -1,25 +1,13 @@
-'use client';
-
-import React from 'react';
-
-import { Banner, NavBar } from '@/components';
-
-import styles from './Homepage.module.scss';
-
-export default function HomePage() {
+import UserProfileCard from '../components/UserLabel';
+export default function Home() {
   return (
-    <div className={styles['homepage-container']}>
-      <div className={styles['banner-background']} />
-
-      <div className={styles['blur-top']} >
-         <NavBar />
-      </div>
-  
-
-      <div className={styles['homepage-content']}>
-       
-        <Banner />
-      </div>
+    <div className='grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]'>
+      <h1>Hello World</h1>
+      {/* Login link(must use <a> link to define Auth0 component) */}
+      <a href="/auth/login">Login</a>
+      {/* Logout link(must use <a> link to define Auth0 component) */}
+      <a href="/auth/logout" >Logout</a>
+      {/* conclusion: The Home component renders a simple layout with login and logout links. */}
     </div>
   );
 }
